SALARIZARE - Ionut Hrubaru - refactor StergeCentralizator in registru

+++ trunk/OpenERP/OpenERP_SALARIZARE/src/org/open/erp/services/salarizare/impl/RegistruSalarizare.java
@@ -226,7 +226,8 @@ public class RegistruSalarizare {
 	}
 	
 	public void stergeCentralizator(CentralizatorStatSalarii centralizator){
-		entityManager.find(centralizator.getClass(), centralizator.getIdCentralizator());
+		centralizator = entityManager.find(centralizator.getClass(), centralizator.getIdCentralizator());
+		if(centralizator!=null)
 		entityManager.remove(centralizator);
 	}
